feat: use explicit casting and converting rules

+++ src/camera/camera.c
@@ -30,7 +30,7 @@ t_camera	*new_camera(int hsize, int vsize, double field_of_view)
 	camera->fov = field_of_view;
 	camera->transform = identity_matrix();
 	camera->inverse_transform = identity_matrix();
-	half_view = tanf(field_of_view / 2);
+	half_view = tan(field_of_view / 2);
 	aspect = (double)hsize / vsize;
 	if (aspect >= 1)
 	{
